requirements.txt
django==3.0
djangorestframework==3.12.0
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
django-heroku==0.3.1
whitenoise==5.2.0
PyYAML==5.4.1
uritemplate==3.0.1
